rename and organize tasks
